fix style, migrate legacy hover serializer
