+++ bom/build.gradle.kts
@@ -23,8 +23,8 @@ dependencies {
       "text-minimessage",
       "text-serializer-ansi",
       "text-serializer-gson",
-      "text-serializer-gson-legacy-impl",
       "text-serializer-json",
+      "text-serializer-json-legacy-impl",
       "text-serializer-legacy",
       "text-serializer-plain"
     ).forEach {
