+++ settings.gradle.kts
@@ -44,6 +44,7 @@ sequenceOf(
   "text-serializer-gson",
   "text-serializer-gson-legacy-impl",
   "text-serializer-json",
+  "text-serializer-json-legacy-impl",
   "text-serializer-legacy",
   "text-serializer-plain",
   "text-serializer-ansi",
