+++ text-serializer-json-legacy-impl/src/main/java/net/kyori/adventure/text/serializer/json/legacyimpl/NBTLegacyHoverEventSerializer.java
@@ -21,31 +21,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.text.serializer.gson.legacyimpl;
+package net.kyori.adventure.text.serializer.json.legacyimpl;
 
 import net.kyori.adventure.text.event.HoverEvent;
-import net.kyori.adventure.text.serializer.gson.LegacyHoverEventSerializer;
-import org.jetbrains.annotations.ApiStatus;
+import net.kyori.adventure.text.serializer.json.LegacyHoverEventSerializer;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * A legacy {@link HoverEvent} serializer.
  *
- * @since 4.3.0
- * @deprecated for removal since 4.14, use {@link net.kyori.adventure.text.serializer.json.legacyimpl.NBTLegacyHoverEventSerializer the text-serializer-json version} instead.
+ * @since 4.14.0
  */
-@Deprecated
-@ApiStatus.ScheduledForRemoval(inVersion = "5.0.0")
-public interface NBTLegacyHoverEventSerializer extends LegacyHoverEventSerializer, net.kyori.adventure.text.serializer.json.legacyimpl.NBTLegacyHoverEventSerializer {
+public interface NBTLegacyHoverEventSerializer extends LegacyHoverEventSerializer {
   /**
    * Gets the legacy {@link HoverEvent} serializer.
    *
    * @return a legacy {@link HoverEvent} serializer
-   * @since 4.3.0
-   * @deprecated for removal since 4.14, use {@link net.kyori.adventure.text.serializer.json.legacyimpl.NBTLegacyHoverEventSerializer the text-serializer-json version} instead.
+   * @since 4.14.0
    */
-  @Deprecated
-  @ApiStatus.ScheduledForRemoval(inVersion = "5.0.0")
   static @NotNull LegacyHoverEventSerializer get() {
     return NBTLegacyHoverEventSerializerImpl.INSTANCE;
   }
